requirements.txt
streamlit
ultralytics
requests
Pillow